Start working through the problem
Add logic for iterating through nums array and checking if value is less
than the next.

Running into issue when value is strictly increasing and needs to skip
to the next value but compare that to the previous value. Will try to
compare values, and if it is less than will add to an array. This will
allow me to check the previous value by recalling the last value in the
confirmed array.